requirements.txt
django==1.9.4
django-storages==1.4.1
boto==2.39.0
Pillow==3.2.0
python-dateutil==2.5.2